Fix for errors detected by GCC 5.1.0.
git-svn-id: svn://svn.code.sf.net/p/sobjectizer/repo/branches/so_5/5.5.8--dev-trunk@1554 1d1731ae-3f85-447d-88f8-c72b288064d1

+++ dev/so_5/disp/prio/common_thread/impl/h/demand_queue.hpp
@@ -209,7 +209,8 @@ class demand_queue_t
 		queue_for_one_priority_t * m_current_priority = nullptr;
 
 		//! Subqueues for priorities.
-		queue_for_one_priority_t m_priorities[ priority_t::p_max ];
+		queue_for_one_priority_t m_priorities[
+				static_cast< std::size_t >( priority_t::p_max ) ];
 
 		//! Destroy all demands in the queue specified.
 		void
